[dpc++][ranges] + moved "the custom ranges stuff" to separate header file.
